construindo combo de 3 ataques -> guerreiro

+++ Assets/Scripts/Nucleo/Interfaces/IComandosGerais.cs
@@ -7,7 +7,7 @@ namespace Assets.Scripts.Nucleo.Interfaces
 	{
 		// Jogadores e inimigos
 		public float ComandoMovimento();//para o método Andar
-		public float ComandoAtaque();//para o método Atacar
+		public int ComandoAtaque();//para o método Atacar
 		public bool ComandoInteracao();//para o método Interagir
 
 		//Jogadores
